Struct octorest::apis::GistsApi

pub struct GistsApi<'t> { /* fields omitted */ }

This is supported on feature="gh-gists" only.

Gists API

Implementations

impl<'client> GistsApi<'client>

pub fn list_public(&self) -> GistsListPublicBuilder

List public gists

List public gists sorted by most recently updated to least recently updated.

Note: With pagination, you can fetch up to 3000 gists. For example, you can fetch 100 pages with 30 gists per page or 30 pages with 100 gists per page.

See also

GitHub Developer Guide

pub fn get_revision(&self) -> GistsGetRevisionBuilder

Get a gist revision

See also

GitHub Developer Guide

pub fn list_for_user(&self) -> GistsListForUserBuilder

List gists for a user

Lists public gists for the specified user:

See also

GitHub Developer Guide

pub fn list(&self) -> GistsListBuilder

List gists for the authenticated user

Lists the authenticated user's gists or if called anonymously, this endpoint returns all public gists:

See also

GitHub Developer Guide

pub fn create(&self) -> GistsCreateBuilder

Create a gist

Allows you to add a new gist with one or more files.

Note: Don't name your files "gistfile" with a numerical suffix. This is the format of the automatic naming scheme that Gist uses internally.

See also

GitHub Developer Guide

pub fn fork(&self) -> GistsForkBuilder

Fork a gist

Note: This was previously /gists/:gist_id/fork.

See also

GitHub Developer Guide

pub fn list_forks(&self) -> GistsListForksBuilder

List gist forks

See also

GitHub Developer Guide

pub fn list_commits(&self) -> GistsListCommitsBuilder

List gist commits

See also

GitHub Developer Guide

pub fn list_starred(&self) -> GistsListStarredBuilder

List starred gists

List the authenticated user's starred gists:

See also

GitHub Developer Guide

pub fn delete(&self) -> GistsDeleteBuilder

Delete a gist

See also

GitHub Developer Guide

pub fn get(&self) -> GistsGetBuilder

Get a gist

See also

GitHub Developer Guide

pub fn update(&self) -> GistsUpdateBuilder

Update a gist

Allows you to update or delete a gist file and rename gist files. Files from the previous version of the gist that aren't explicitly changed during an edit are unchanged.

See also

GitHub Developer Guide

pub fn get_comment(&self) -> GistsGetCommentBuilder

Get a gist comment

See also

GitHub Developer Guide

pub fn delete_comment(&self) -> GistsDeleteCommentBuilder

Delete a gist comment

See also

GitHub Developer Guide

pub fn update_comment(&self) -> GistsUpdateCommentBuilder

Update a gist comment

See also

GitHub Developer Guide

pub fn list_comments(&self) -> GistsListCommentsBuilder

List gist comments

See also

GitHub Developer Guide

pub fn create_comment(&self) -> GistsCreateCommentBuilder

Create a gist comment

See also

GitHub Developer Guide

pub fn star(&self) -> GistsStarBuilder

Star a gist

Note that you'll need to set Content-Length to zero when calling out to this endpoint. For more information, see "HTTP verbs."

See also

GitHub Developer Guide

pub fn unstar(&self) -> GistsUnstarBuilder

Unstar a gist

See also

GitHub Developer Guide

pub fn check_is_starred(&self) -> GistsCheckIsStarredBuilder

Check if a gist is starred

See also

GitHub Developer Guide