Struct octorest::apis::IssuesApi

pub struct IssuesApi<'t> { /* fields omitted */ }

This is supported on feature="gh-issues" only.

Issues API

Implementations

impl<'client> IssuesApi<'client>

pub fn list_events_for_timeline(&self) -> IssuesListEventsForTimelineBuilder

List timeline events for an issue

See also

GitHub Developer Guide

pub fn check_user_can_be_assigned(&self) -> IssuesCheckUserCanBeAssignedBuilder

Check if a user can be assigned

Checks if a user has permission to be assigned to an issue in this repository.

If the assignee can be assigned to issues in the repository, a 204 header with no content is returned.

Otherwise a 404 status code is returned.

See also

GitHub Developer Guide

pub fn list_labels_for_repo(&self) -> IssuesListLabelsForRepoBuilder

List labels for a repository

See also

GitHub Developer Guide

pub fn create_label(&self) -> IssuesCreateLabelBuilder

Create a label

See also

GitHub Developer Guide

pub fn lock(&self) -> IssuesLockBuilder

Lock an issue

Users with push access can lock an issue or pull request's conversation.

Note that, if you choose not to pass any parameters, you'll need to set Content-Length to zero when calling out to this endpoint. For more information, see "HTTP verbs."

See also

GitHub Developer Guide

pub fn unlock(&self) -> IssuesUnlockBuilder

Unlock an issue

Users with push access can unlock an issue's conversation.

See also

GitHub Developer Guide

pub fn create(&self) -> IssuesCreateBuilder

Create an issue

Any user with pull access to a repository can create an issue. If issues are disabled in the repository, the API returns a 410 Gone status.

This endpoint triggers notifications. Creating content too quickly using this endpoint may result in abuse rate limiting. See "Abuse rate limits" and "Dealing with abuse rate limits" for details.

See also

GitHub Developer Guide

pub fn list_for_repo(&self) -> IssuesListForRepoBuilder

List repository issues

List issues in a repository.

Note: GitHub's REST API v3 considers every pull request an issue, but not every issue is a pull request. For this reason, "Issues" endpoints may return both issues and pull requests in the response. You can identify pull requests by the pull_request key.

Be aware that the id of a pull request returned from "Issues" endpoints will be an issue id. To find out the pull request id, use the "List pull requests" endpoint.

See also

GitHub Developer Guide

pub fn list_for_org(&self) -> IssuesListForOrgBuilder

List organization issues assigned to the authenticated user

List issues in an organization assigned to the authenticated user.

Note: GitHub's REST API v3 considers every pull request an issue, but not every issue is a pull request. For this reason, "Issues" endpoints may return both issues and pull requests in the response. You can identify pull requests by the pull_request key.

Be aware that the id of a pull request returned from "Issues" endpoints will be an issue id. To find out the pull request id, use the "List pull requests" endpoint.

See also

GitHub Developer Guide

pub fn get_event(&self) -> IssuesGetEventBuilder

Get an issue event

See also

GitHub Developer Guide

pub fn list_for_authenticated_user(
&self
) -> IssuesListForAuthenticatedUserBuilder

List user account issues assigned to the authenticated user

List issues across owned and member repositories assigned to the authenticated user:

Note: GitHub's REST API v3 considers every pull request an issue, but not every issue is a pull request. For this reason, "Issues" endpoints may return both issues and pull requests in the response. You can identify pull requests by the pull_request key.

Be aware that the id of a pull request returned from "Issues" endpoints will be an issue id. To find out the pull request id, use the "List pull requests" endpoint.

See also

GitHub Developer Guide

pub fn remove_label(&self) -> IssuesRemoveLabelBuilder

Remove a label from an issue

Removes the specified label from the issue, and returns the remaining labels on the issue. This endpoint returns a 404 Not Found status if the label does not exist.

See also

GitHub Developer Guide

pub fn update_comment(&self) -> IssuesUpdateCommentBuilder

Update an issue comment

See also

GitHub Developer Guide

pub fn delete_comment(&self) -> IssuesDeleteCommentBuilder

Delete an issue comment

See also

GitHub Developer Guide

pub fn get_comment(&self) -> IssuesGetCommentBuilder

Get an issue comment

See also

GitHub Developer Guide

pub fn list_events_for_repo(&self) -> IssuesListEventsForRepoBuilder

List issue events for a repository

See also

GitHub Developer Guide

pub fn list(&self) -> IssuesListBuilder

List issues assigned to the authenticated user

List issues assigned to the authenticated user across all visible repositories including owned repositories, member repositories, and organization repositories. You can use the filter query parameter to fetch issues that are not necessarily assigned to you. See the Parameters table for more information.

Note: GitHub's REST API v3 considers every pull request an issue, but not every issue is a pull request. For this reason, "Issues" endpoints may return both issues and pull requests in the response. You can identify pull requests by the pull_request key.

Be aware that the id of a pull request returned from "Issues" endpoints will be an issue id. To find out the pull request id, use the "List pull requests" endpoint.

See also

GitHub Developer Guide

pub fn update(&self) -> IssuesUpdateBuilder

Update an issue

Issue owners and users with push access can edit an issue.

See also

GitHub Developer Guide

pub fn get(&self) -> IssuesGetBuilder

Get an issue

The API returns a 301 Moved Permanently status if the issue was transferred to another repository. If the issue was transferred to or deleted from a repository where the authenticated user lacks read access, the API returns a 404 Not Found status. If the issue was deleted from a repository where the authenticated user has read access, the API returns a 410 Gone status. To receive webhook events for transferred and deleted issues, subscribe to the issues webhook.

Note: GitHub's REST API v3 considers every pull request an issue, but not every issue is a pull request. For this reason, "Issues" endpoints may return both issues and pull requests in the response. You can identify pull requests by the pull_request key.

Be aware that the id of a pull request returned from "Issues" endpoints will be an issue id. To find out the pull request id, use the "List pull requests" endpoint.

See also

GitHub Developer Guide

pub fn create_comment(&self) -> IssuesCreateCommentBuilder

Create an issue comment

This endpoint triggers notifications. Creating content too quickly using this endpoint may result in abuse rate limiting. See "Abuse rate limits" and "Dealing with abuse rate limits" for details.

See also

GitHub Developer Guide

pub fn list_comments(&self) -> IssuesListCommentsBuilder

List issue comments

Issue Comments are ordered by ascending ID.

See also

GitHub Developer Guide

pub fn list_milestones(&self) -> IssuesListMilestonesBuilder

List milestones

See also

GitHub Developer Guide

pub fn create_milestone(&self) -> IssuesCreateMilestoneBuilder

Create a milestone

See also

GitHub Developer Guide

pub fn list_labels_on_issue(&self) -> IssuesListLabelsOnIssueBuilder

List labels for an issue

See also

GitHub Developer Guide

pub fn add_labels(&self) -> IssuesAddLabelsBuilder

Add labels to an issue

See also

GitHub Developer Guide

pub fn set_labels(&self) -> IssuesSetLabelsBuilder

Set labels for an issue

Removes any previous labels and sets the new labels for an issue.

See also

GitHub Developer Guide

pub fn remove_all_labels(&self) -> IssuesRemoveAllLabelsBuilder

Remove all labels from an issue

See also

GitHub Developer Guide

pub fn update_milestone(&self) -> IssuesUpdateMilestoneBuilder

Update a milestone

See also

GitHub Developer Guide

pub fn delete_milestone(&self) -> IssuesDeleteMilestoneBuilder

Delete a milestone

See also

GitHub Developer Guide

pub fn get_milestone(&self) -> IssuesGetMilestoneBuilder

Get a milestone

See also

GitHub Developer Guide

pub fn list_events(&self) -> IssuesListEventsBuilder

List issue events

See also

GitHub Developer Guide

pub fn list_labels_for_milestone(&self) -> IssuesListLabelsForMilestoneBuilder

List labels for issues in a milestone

See also

GitHub Developer Guide

pub fn add_assignees(&self) -> IssuesAddAssigneesBuilder

Add assignees to an issue

Adds up to 10 assignees to an issue. Users already assigned to an issue are not replaced.

This example adds two assignees to the existing octocat assignee.

See also

GitHub Developer Guide

pub fn remove_assignees(&self) -> IssuesRemoveAssigneesBuilder

Remove assignees from an issue

Removes one or more assignees from an issue.

This example removes two of three assignees, leaving the octocat assignee.

See also

GitHub Developer Guide

pub fn list_comments_for_repo(&self) -> IssuesListCommentsForRepoBuilder

List issue comments for a repository

By default, Issue Comments are ordered by ascending ID.

See also

GitHub Developer Guide

pub fn delete_label(&self) -> IssuesDeleteLabelBuilder

Delete a label

See also

GitHub Developer Guide

pub fn get_label(&self) -> IssuesGetLabelBuilder

Get a label

See also

GitHub Developer Guide

pub fn update_label(&self) -> IssuesUpdateLabelBuilder

Update a label

See also

GitHub Developer Guide

pub fn list_assignees(&self) -> IssuesListAssigneesBuilder

List assignees

Lists the available assignees for issues in a repository.

See also

GitHub Developer Guide