Struct octorest::apis::ReactionsApi

pub struct ReactionsApi<'t> { /* fields omitted */ }

This is supported on feature="gh-reactions" only.

Reactions API

Implementations

impl<'client> ReactionsApi<'client>

pub fn delete_for_commit_comment(
&self
) -> ReactionsDeleteForCommitCommentBuilder

Delete a commit comment reaction

Note: You can also specify a repository by repository_id using the route DELETE /repositories/:repository_id/comments/:comment_id/reactions/:reaction_id.

Delete a reaction to a commit comment.

See also

GitHub Developer Guide

pub fn list_for_issue_comment(&self) -> ReactionsListForIssueCommentBuilder

List reactions for an issue comment

List the reactions to an issue comment.

See also

GitHub Developer Guide

pub fn create_for_issue_comment(&self) -> ReactionsCreateForIssueCommentBuilder

Create reaction for an issue comment

Create a reaction to an issue comment. A response with a Status: 200 OK means that you already added the reaction type to this issue comment.

See also

GitHub Developer Guide

pub fn delete_for_issue(&self) -> ReactionsDeleteForIssueBuilder

Delete an issue reaction

Note: You can also specify a repository by repository_id using the route DELETE /repositories/:repository_id/issues/:issue_number/reactions/:reaction_id.

Delete a reaction to an issue.

See also

GitHub Developer Guide

pub fn delete_for_team_discussion(
&self
) -> ReactionsDeleteForTeamDiscussionBuilder

Delete team discussion reaction

Note: You can also specify a team or organization with team_id and org_id using the route DELETE /organizations/:org_id/team/:team_id/discussions/:discussion_number/reactions/:reaction_id.

Delete a reaction to a team discussion. OAuth access tokens require the write:discussion scope.

See also

GitHub Developer Guide

pub fn delete_for_issue_comment(&self) -> ReactionsDeleteForIssueCommentBuilder

Delete an issue comment reaction

Note: You can also specify a repository by repository_id using the route DELETE delete /repositories/:repository_id/issues/comments/:comment_id/reactions/:reaction_id.

Delete a reaction to an issue comment.

See also

GitHub Developer Guide

pub fn create_for_team_discussion_in_org(
&self
) -> ReactionsCreateForTeamDiscussionInOrgBuilder

Create reaction for a team discussion

Create a reaction to a team discussion. OAuth access tokens require the write:discussion scope. A response with a Status: 200 OK means that you already added the reaction type to this team discussion.

Note: You can also specify a team by org_id and team_id using the route POST /organizations/:org_id/team/:team_id/discussions/:discussion_number/reactions.

See also

GitHub Developer Guide

pub fn list_for_team_discussion_in_org(
&self
) -> ReactionsListForTeamDiscussionInOrgBuilder

List reactions for a team discussion

List the reactions to a team discussion. OAuth access tokens require the read:discussion scope.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id/discussions/:discussion_number/reactions.

See also

GitHub Developer Guide

pub fn delete_legacy(&self) -> ReactionsDeleteLegacyBuilder

Delete a reaction (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Reactions API. We recommend migrating your existing code to use the new delete reactions endpoints. For more information, see this blog post.

OAuth access tokens require the write:discussion scope, when deleting a team discussion or team discussion comment.

See also

GitHub Developer Guide

pub fn list_for_team_discussion_legacy(
&self
) -> ReactionsListForTeamDiscussionLegacyBuilder

List reactions for a team discussion (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new List reactions for a team discussion endpoint.

List the reactions to a team discussion. OAuth access tokens require the read:discussion scope.

See also

GitHub Developer Guide

pub fn create_for_team_discussion_legacy(
&self
) -> ReactionsCreateForTeamDiscussionLegacyBuilder

Create reaction for a team discussion (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Create reaction for a team discussion endpoint.

Create a reaction to a team discussion. OAuth access tokens require the write:discussion scope. A response with a Status: 200 OK means that you already added the reaction type to this team discussion.

See also

GitHub Developer Guide

pub fn delete_for_pull_request_comment(
&self
) -> ReactionsDeleteForPullRequestCommentBuilder

Delete a pull request comment reaction

Note: You can also specify a repository by repository_id using the route DELETE /repositories/:repository_id/pulls/comments/:comment_id/reactions/:reaction_id.

Delete a reaction to a pull request review comment.

See also

GitHub Developer Guide

pub fn list_for_team_discussion_comment_in_org(
&self
) -> ReactionsListForTeamDiscussionCommentInOrgBuilder

List reactions for a team discussion comment

List the reactions to a team discussion comment. OAuth access tokens require the read:discussion scope.

Note: You can also specify a team by org_id and team_id using the route GET /organizations/:org_id/team/:team_id/discussions/:discussion_number/comments/:comment_number/reactions.

See also

GitHub Developer Guide

pub fn create_for_team_discussion_comment_in_org(
&self
) -> ReactionsCreateForTeamDiscussionCommentInOrgBuilder

Create reaction for a team discussion comment

Create a reaction to a team discussion comment. OAuth access tokens require the write:discussion scope. A response with a Status: 200 OK means that you already added the reaction type to this team discussion comment.

Note: You can also specify a team by org_id and team_id using the route POST /organizations/:org_id/team/:team_id/discussions/:discussion_number/comments/:comment_number/reactions.

See also

GitHub Developer Guide

pub fn create_for_commit_comment(
&self
) -> ReactionsCreateForCommitCommentBuilder

Create reaction for a commit comment

Create a reaction to a commit comment. A response with a Status: 200 OK means that you already added the reaction type to this commit comment.

See also

GitHub Developer Guide

pub fn list_for_commit_comment(&self) -> ReactionsListForCommitCommentBuilder

List reactions for a commit comment

List the reactions to a commit comment.

See also

GitHub Developer Guide

pub fn delete_for_team_discussion_comment(
&self
) -> ReactionsDeleteForTeamDiscussionCommentBuilder

Delete team discussion comment reaction

Note: You can also specify a team or organization with team_id and org_id using the route DELETE /organizations/:org_id/team/:team_id/discussions/:discussion_number/comments/:comment_number/reactions/:reaction_id.

Delete a reaction to a team discussion comment. OAuth access tokens require the write:discussion scope.

See also

GitHub Developer Guide

pub fn create_for_pull_request_review_comment(
&self
) -> ReactionsCreateForPullRequestReviewCommentBuilder

Create reaction for a pull request review comment

Create a reaction to a pull request review comment. A response with a Status: 200 OK means that you already added the reaction type to this pull request review comment.

See also

GitHub Developer Guide

pub fn list_for_pull_request_review_comment(
&self
) -> ReactionsListForPullRequestReviewCommentBuilder

List reactions for a pull request review comment

List the reactions to a pull request review comment.

See also

GitHub Developer Guide

pub fn list_for_team_discussion_comment_legacy(
&self
) -> ReactionsListForTeamDiscussionCommentLegacyBuilder

List reactions for a team discussion comment (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new List reactions for a team discussion comment endpoint.

List the reactions to a team discussion comment. OAuth access tokens require the read:discussion scope.

See also

GitHub Developer Guide

pub fn create_for_team_discussion_comment_legacy(
&self
) -> ReactionsCreateForTeamDiscussionCommentLegacyBuilder

Create reaction for a team discussion comment (Legacy)

Deprecation Notice: This endpoint route is deprecated and will be removed from the Teams API. We recommend migrating your existing code to use the new Create reaction for a team discussion comment endpoint.

Create a reaction to a team discussion comment. OAuth access tokens require the write:discussion scope. A response with a Status: 200 OK means that you already added the reaction type to this team discussion comment.

See also

GitHub Developer Guide

pub fn create_for_issue(&self) -> ReactionsCreateForIssueBuilder

Create reaction for an issue

Create a reaction to an issue. A response with a Status: 200 OK means that you already added the reaction type to this issue.

See also

GitHub Developer Guide

pub fn list_for_issue(&self) -> ReactionsListForIssueBuilder

List reactions for an issue

List the reactions to an issue.

See also

GitHub Developer Guide